requirements.txt
anthropic>=0.40.0
openai>=1.50.0
scipy>=1.12.0
tabulate>=0.9.0
tqdm>=4.66.0
pyyaml>=6.0
